Set `serverless-http` and `@aws-sdk/client-s3` as peerDependencies.
These two packages are not specifically required to run open-next.

However, they are required in order for the lambda functionc to be
bundled correctly. Therefore, it makes sense to require them to be
installed as dependencies of the Next.js application we are trying to
bundle and not as dependencies of open-next itself.